refactor: move model initializer to a private Database util
